refactor: merge login and signup Fixes #39
